web: * web初始化失败后自动回滚操作 * 检查数据线bug修复 * 修复kafka reader页面卡顿的bug * zookeeper部分模板修改 * 拓扑stopped下，表任意状态都可以删除 * 多租户namespace修改,在instance后添加router名称用!分隔 * 资源管理tab页，已选择列表带过滤功能 * 优化资源添加:添加一键添加所有表功能 * 全量拉取记录添加编辑功能 * 一键清除全量报警(数据线级别) * 登录界面添加账户名为邮箱的提示，防止误解 全量: * 全量拉取主键为bigint的表数据条数对不上的bug修复(bugbigint转成int丢失精度导致) * 全量mysql脱敏列大小写bug修复(查询语句列名转成了大写)
* canal-auto-deploy  bugfix

+++ dbus-fullpuller/src/main/java/com/creditease/dbus/common/splitters/FloatSplitter.java
@@ -61,7 +61,8 @@ public class FloatSplitter implements DBSplitter  {
 
   private static final double MIN_INCREMENT = 10000 * Double.MIN_VALUE;
 
-  private final int type = Types.FLOAT;
+  //private final int type = Types.FLOAT;
+  private final int type = Types.DOUBLE;
 
   public List<InputSplit> split(long numSplits, ResultSet results,
       String colName, DBConfiguration dbConf) throws SQLException {
