+++ dbus-keeper/keeper-base/src/main/java/com/creditease/dbus/domain/model/TableVersion.java
@@ -34,8 +34,8 @@ public class TableVersion {
     private String tableName;
     private Integer version;
     private Integer innerVersion;
-    private Integer eventOffset;
-    private Integer eventPos;
+    private Long eventOffset;
+    private Long eventPos;
     private Date updateTime;
     private String comments;
 
@@ -103,19 +103,19 @@ public class TableVersion {
         this.innerVersion = innerVersion;
     }
 
-    public Integer getEventOffset() {
+    public Long getEventOffset() {
         return eventOffset;
     }
 
-    public void setEventOffset(Integer eventOffset) {
+    public void setEventOffset(Long eventOffset) {
         this.eventOffset = eventOffset;
     }
 
-    public Integer getEventPos() {
+    public Long getEventPos() {
         return eventPos;
     }
 
-    public void setEventPos(Integer eventPos) {
+    public void setEventPos(Long eventPos) {
         this.eventPos = eventPos;
     }
